Fix: Se arreglo el bug del eje Y

+++ server_src/posicion.cpp
@@ -26,10 +26,6 @@ bool Posicion::mover(const std::string& direccion) {
     return false;
 }
 
-int Posicion::get_posicion_x() {
-    return this->x;
-}
+uint16_t Posicion::get_posicion_x() { return this->x; }
 
-int Posicion::get_posicion_y() {
-    return this->y;
-}
+uint16_t Posicion::get_posicion_y() { return this->y; }
